+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-15T16:01:51.772Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-16T06:43:09.103Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")