+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7M9PT2E1J0PEMVHBCJ1MNKM"
+        "01K7NW4EM9HGV9NDCGDREGGTG5"
     )
     r"""The id of the resource"""