+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-15T16:01:55.517Z")
+    created: Optional[datetime] = parse_datetime("2025-10-16T06:43:12.275Z")
     r"""The date and time the resource was created"""