+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7M9PT2D6KB4C185W8GCTH0X"
+        "01K7NW4EM6FM2FYHMM70GPYPR0"
     )
     r"""The id of the resource"""