+++ packages/orq-rc/RELEASES.md
@@ -3349,3 +3349,13 @@ Based on:
 - [python v3.14.0-rc.30] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.30] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.30 - packages/orq-rc
+
+## 2025-10-16 06:44:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.31] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.31] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.31 - packages/orq-rc